services/auth-service/requirements.txt
flask
flask-cors
python-dotenv
google-auth
google-auth-oauthlib
